[quant][graphmode] Minor fix and turn off fold_convbn
Summary:
In fold_convbn pass, we need to recompute the parameter(weight, bias) for
conv, update the attribute of conv and update the access of bias in conv
because if the original conv have no bias, the `self.bias` access will be
inline and replaced by Constant node `None = prim::Constant()`, we need to
update this to use `GetAttr[name="bias"]` to make this work. But there is
also some work going on the handle constants, so we'll fix this pass after
that is done.

Test Plan:
.

Reviewers:
mvz

Subscribers:

Tasks:

Tags:

ghstack-source-id: b409056da2cb75c7dbcb6b6ef0ff5591802950bb
Pull Request resolved: https://github.com/pytorch/pytorch/pull/27403