Add CallStack class.
Callstack serves a similar purpose to Scope, but instead of storing
string names of the functions it stores pointer to Function objects
themselves. Currently, scopes are used in tracing and callstacks are
used in scripting - hopefully I would be able to merge them in future.

ghstack-source-id: 227b1b4ba6feaff45de0c541f8d4254aeb8be7e8

gh-metadata: pytorch pytorch 27917 gh/ZolotukhinM/135/head