[pytorch] Double-buffer in Pickler to reduce #calls.
This change adds a small fixed-size buffer to Pickler to
avoid calling writer_() and the associated downstream checks
on a per-opcode/per-byte basis.

We end up still doing a bounds check in the common case,
but the memcpy() is a fixed size. And we reduce the number
of backend calls.

In practice, this change speeds up the Pickle1MInts benchmark
for me locally from roughly 56msec to 22msec.

Additionally, in this change we convert a few pushIValue() on
typed lists, where we know the type to be double/int/boot to be
pushInt() to bypass a bit of logic.

We should additionally change the Unpickler, though keeping
this separate, since the std::function<> prototype needs to be
changed to do buffering (return value needs to change from bool
to size_t)

Differential Revision: [D17939311](https://our.internmc.facebook.com/intern/diff/D17939311/)

ghstack-source-id: 91964964
Pull Request resolved: https://github.com/pytorch/pytorch/pull/28043