[pytorch] When pickling specialized lists, reduce prior to adding values.
Previously, we weren't converting a generic list to specialized (int/double/etc)
until after we read all the values.

If we slightly change the order of the pickler opcodes, we can convert the
list to specialized before we read the values. And if the unpickler knows the
specialized type when reading, it can skip the per-entry IValue
constructor/destructor step.

This produces a 45% win for this specialized case - in my benchmark unpickling
a 1M random int list, the time goes from 65ms -> 36ms (both are still slow, but it's better).

Differential Revision: [D18078506](https://our.internmc.facebook.com/intern/diff/D18078506/)

ghstack-source-id: 92434621
Pull Request resolved: https://github.com/pytorch/pytorch/pull/28488