[jit] fix @property
Currently `@property` methods are broken in TorchScript because we
basically treat it as an attribute in the existing path: we'll evaluate
the method once and store that as the value forever.

Since lack of `@property` support is easily worked around (just make it
a method), I've opted to just explicitly error to avoid confusion. If
people want it, they can file an issue and we can look at their use
case.

This also helps us nicely clean up some parts of the ScriptModule conversion
path.

ghstack-source-id: d318663429370b23a12ec3a7e72bdf1417d03a1d
Pull Request resolved: https://github.com/pytorch/pytorch/pull/28395