[lint] make sure clang-tidy is diffing against the right thing
Okay, my last fix was wrong because it turns out that the base SHA is
computed at PR time using the actual repo's view of the base ref, not
the user's. So if the user doesn't rebase on top of the latest master
before putting up the PR, the diff thing is wrong anyway.

This PR fixes the issue by not relying on any of these API details and
just getting the merge-base of the base and head refs, which should
guarantee we are diffing against the right thing.

ghstack-source-id: 9a14155b5dbe580cc47fecc0bce72b42a12cf3ca
Pull Request resolved: https://github.com/pytorch/pytorch/pull/28796