[namedtensor] add support for {ones,zeros,full,rand,randn}_like ops
This PR adds support for calling those functions on named tensors. The
implementation is not the nicest: in the future we have plans to merge
names into TensorOptions, at which point we don't need the extra
branches that check if the tensor has names. Right now, however, these
functions are very useful to have (in particular, ones_like is used by
autograd to generate gradients).

Test Plan:
- Added tests for each of these

ghstack-source-id: 65a94e20e152304ea2b855a9f6208fefb0ea6f63
Pull Request resolved: https://github.com/pytorch/pytorch/pull/28981